+++ db/migrate/20230901090637_add_supporting_documents_reminder_sent_at_to_users.rb
@@ -1,0 +1,5 @@
+class AddSupportingDocumentsReminderSentAtToUsers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :supporting_documents_reminder_sent_at, :datetime
+  end
+end